requirements.txt
aiogram==3.12.0
Pillow==10.2.0
sympy==1.12
